# Conflicts:
#	Cargo.lock

+++ hopr/hopr-lib/src/lib.rs
@@ -369,7 +369,7 @@ where
         self.error_if_not_in_state(HoprState::Running, "Node is not ready for on-chain operations".into())?;
 
         let backoff = backon::ConstantBuilder::default()
-            .with_max_times(self.cfg.protocol.session.establish_max_retries as usize)
+            .with_max_times(self.cfg.protocol.session.establish_max_retries)
             .with_delay(self.cfg.protocol.session.establish_retry_timeout)
             .with_jitter();
 
@@ -479,6 +479,8 @@ where
         F: Fn(&ChainEvent) -> bool + Send + Sync + 'static,
     {
         debug!(%context, "registering wait for on-chain event");
+
+        // DropAbortable not needed because the stream only generates items when polled
         let (event_stream, handle) = futures::stream::abortable(
             self.chain_api
                 .subscribe()?
@@ -750,9 +752,6 @@ pub fn peer_id_to_offchain_key(peer_id: &PeerId) -> errors::Result<OffchainPubli
 
 #[cfg(test)]
 mod tests {
-    use anyhow::Context as _;
-    use hopr_transport::Keypair;
-
     use super::*;
 
     #[test]
@@ -866,6 +865,9 @@ mod tests {
     #[allow(deprecated)]
     #[test]
     fn explicit_path_config_converts_into_intermediate_path_routing_options() -> anyhow::Result<()> {
+        use anyhow::Context as _;
+        use hopr_transport::Keypair;
+
         let k1 = hopr_api::types::internal::NodeId::from(*OffchainKeypair::random().public());
         let k2 = hopr_api::types::internal::NodeId::from(*OffchainKeypair::random().public());
         let k3 = hopr_api::types::internal::NodeId::from(*OffchainKeypair::random().public());
